added images, used blur
