+++ messages/np.json
@@ -1,7 +1,8 @@
 {
   "Translations": {
     "logoQuiz": {
-      "finishQuiz": "क्विज समाप्त गर्नुहोस्",
+    "proTip": "प्रो टिप: हरेक गलत अनुमानपछि तस्बिर अझ प्रष्ट हुँदै जान्छ, जसले चिन्न सजिलो हुन्छ!",
+    "finishQuiz": "क्विज समाप्त गर्नुहोस्",
     "pageTitle": "नेपाल लोगो क्विज",
     "pageDescription": "लोकप्रिय नेपाली र अन्तर्राष्ट्रिय ब्रान्ड लोगोहरूको तपाईंको ज्ञान परीक्षण गर्नुहोस्। तपाईं कति चिन्न सक्नुहुन्छ?",
     "title": "लोगो क्विज",
