+++ src/app/data/logo-quiz/getLogos.js
@@ -6,7 +6,7 @@ const logoData = {
       {
         id: 'ncell',
         name: 'Ncell',
-        imagePath: '/logos/ncell.png',
+        imagePath: '/yo-chineu/ncell.png',
         difficulty: 'easy',
         category: 'telecom',
         acceptableAnswers: ['Ncell', 'N-Cell']
@@ -14,7 +14,7 @@ const logoData = {
       {
         id: 'nabil_bank',
         name: 'Nabil Bank',
-        imagePath: '/logos/nabil.png',
+        imagePath: '/yo-chineu/nabil-bank.png',
         difficulty: 'medium',
         category: 'banking',
         acceptableAnswers: ['Nabil Bank', 'Nabil', 'NABIL']
@@ -22,7 +22,7 @@ const logoData = {
       {
         id: 'buddha_air',
         name: 'Buddha Air',
-        imagePath: '/yo-chineu/buddha_air.png',
+        imagePath: '/yo-chineu/buddha-air.png',
         difficulty: 'medium',
         category: 'airline',
         acceptableAnswers: ['Buddha Air', 'Buddha Airlines']
@@ -30,7 +30,7 @@ const logoData = {
       {
         id: 'tata',
         name: 'Tata',
-        imagePath: '/logos/tata.png',
+        imagePath: '/yo-chineu/tata.jpg',
         difficulty: 'easy',
         category: 'automotive',
         acceptableAnswers: ['Tata', 'Tata Motors', 'TATA']
@@ -38,7 +38,7 @@ const logoData = {
       {
         id: 'wai_wai',
         name: 'Wai Wai',
-        imagePath: '/logos/wai_wai.png',
+        imagePath: '/yo-chineu/wai-wai.png',
         difficulty: 'easy',
         category: 'food',
         acceptableAnswers: ['Wai Wai', 'WaiWai', 'Wai-Wai']
@@ -46,7 +46,7 @@ const logoData = {
       {
         id: 'cg',
         name: 'CG',
-        imagePath: '/logos/cg.png',
+        imagePath: '/yo-chineu/cg.png',
         difficulty: 'medium',
         category: 'conglomerate',
         acceptableAnswers: ['CG', 'Chaudhary Group', 'CG Group']
@@ -54,7 +54,7 @@ const logoData = {
       {
         id: 'dish_home',
         name: 'Dish Home',
-        imagePath: '/logos/dish_home.png',
+        imagePath: '/yo-chineu/dish-home.png',
         difficulty: 'medium',
         category: 'cable',
         acceptableAnswers: ['Dish Home', 'DishHome']
@@ -62,7 +62,7 @@ const logoData = {
       {
         id: 'ntc',
         name: 'Nepal Telecom',
-        imagePath: '/logos/ntc.png',
+        imagePath: '/yo-chineu/ntc.jpg',
         difficulty: 'easy',
         category: 'telecom',
         acceptableAnswers: ['Nepal Telecom', 'NTC', 'Nepal Doorsanchar']
@@ -70,7 +70,7 @@ const logoData = {
       {
         id: 'himalayan_bank',
         name: 'Himalayan Bank',
-        imagePath: '/logos/himalayan_bank.png',
+        imagePath: '/yo-chineu/hbl.png',
         difficulty: 'medium',
         category: 'banking',
         acceptableAnswers: ['Himalayan Bank', 'HBL', 'Himalayan Bank Limited']
@@ -78,7 +78,7 @@ const logoData = {
       {
         id: 'yeti_airlines',
         name: 'Yeti Airlines',
-        imagePath: '/logos/yeti_airlines.png',
+        imagePath: '/yo-chineu/yeti-air.png',
         difficulty: 'medium',
         category: 'airline',
         acceptableAnswers: ['Yeti Airlines', 'Yeti']
@@ -86,7 +86,7 @@ const logoData = {
       {
         id: 'kumari_bank',
         name: 'Kumari Bank',
-        imagePath: '/logos/kumari_bank.png',
+        imagePath: '/yo-chineu/kumari-bank.png',
         difficulty: 'hard',
         category: 'banking',
         acceptableAnswers: ['Kumari Bank', 'Kumari', 'Kumari Bank Limited']
@@ -94,7 +94,7 @@ const logoData = {
       {
         id: 'laxmi_bank',
         name: 'Laxmi Bank',
-        imagePath: '/logos/laxmi_bank.png',
+        imagePath: '/yo-chineu/laxmi-bank.jpg',
         difficulty: 'hard',
         category: 'banking',
         acceptableAnswers: ['Laxmi Bank', 'Laxmi', 'Lakshmi Bank']
@@ -102,7 +102,7 @@ const logoData = {
       {
         id: 'nic_asia',
         name: 'NIC Asia Bank',
-        imagePath: '/logos/nic_asia.png',
+        imagePath: '/yo-chineu/nic-asia.png',
         difficulty: 'hard',
         category: 'banking',
         acceptableAnswers: ['NIC Asia', 'NIC Asia Bank', 'NIC']
@@ -110,7 +110,7 @@ const logoData = {
       {
         id: 'nepal_investment_bank',
         name: 'Nepal Investment Bank',
-        imagePath: '/logos/nepal_investment_bank.png',
+        imagePath: '/yo-chineu/nimb-bank.png',
         difficulty: 'hard',
         category: 'banking',
         acceptableAnswers: ['Nepal Investment Bank', 'NIBL', 'Nepal Investment']
@@ -118,23 +118,15 @@ const logoData = {
       {
         id: 'gorkha_brewery',
         name: 'Gorkha Brewery',
-        imagePath: '/logos/gorkha_brewery.png',
+        imagePath: '/yo-chineu/gorkha-brewery.png',
         difficulty: 'medium',
         category: 'beverage',
         acceptableAnswers: ['Gorkha Brewery', 'Gorkha', 'Gorkha Beer']
       },
       {
-        id: 'nepal_airlines',
-        name: 'Nepal Airlines',
-        imagePath: '/logos/nepal_airlines.png',
-        difficulty: 'easy',
-        category: 'airline',
-        acceptableAnswers: ['Nepal Airlines', 'NAC', 'Nepal Airlines Corporation', 'Royal Nepal Airlines']
-      },
-      {
         id: 'esewa',
         name: 'eSewa',
-        imagePath: '/logos/esewa.png',
+        imagePath: '/yo-chineu/esewa.jpg',
         difficulty: 'easy',
         category: 'fintech',
         acceptableAnswers: ['eSewa', 'e-Sewa', 'E Sewa']
@@ -142,7 +134,7 @@ const logoData = {
       {
         id: 'khalti',
         name: 'Khalti',
-        imagePath: '/logos/khalti.png',
+        imagePath: '/yo-chineu/khalti.png',
         difficulty: 'medium',
         category: 'fintech',
         acceptableAnswers: ['Khalti', 'Khalti Digital Wallet']
@@ -150,7 +142,7 @@ const logoData = {
       {
         id: 'worldlink',
         name: 'WorldLink',
-        imagePath: '/logos/worldlink.png',
+        imagePath: '/yo-chineu/worldlink.png',
         difficulty: 'easy',
         category: 'internet',
         acceptableAnswers: ['WorldLink', 'World Link', 'Worldlink Communications']
@@ -158,7 +150,7 @@ const logoData = {
       {
         id: 'siddhartha_bank',
         name: 'Siddhartha Bank',
-        imagePath: '/logos/siddhartha_bank.png',
+        imagePath: '/yo-chineu/siddhartha-bank.jpg',
         difficulty: 'hard',
         category: 'banking',
         acceptableAnswers: ['Siddhartha Bank', 'Siddhartha', 'Siddhartha Bank Limited']
@@ -168,7 +160,7 @@ const logoData = {
       {
         id: 'ncell',
         name: 'एनसेल',
-        imagePath: '/logos/ncell.png',
+        imagePath: '/yo-chineu/ncell.png',
         difficulty: 'easy',
         category: 'दूरसञ्चार',
         acceptableAnswers: ['एनसेल', 'Ncell', 'N-Cell']
@@ -176,7 +168,7 @@ const logoData = {
       {
         id: 'nabil_bank',
         name: 'नबिल बैंक',
-        imagePath: '/logos/nabil.png',
+        imagePath: '/yo-chineu/nabil.png',
         difficulty: 'medium',
         category: 'बैंकिङ',
         acceptableAnswers: ['नबिल बैंक', 'नबिल', 'Nabil Bank', 'NABIL']
@@ -184,7 +176,7 @@ const logoData = {
       {
         id: 'buddha_air',
         name: 'बुद्ध एयर',
-        imagePath: '/logos/buddha_air.png',
+        imagePath: '/yo-chineu/buddha_air.png',
         difficulty: 'medium',
         category: 'हवाई सेवा',
         acceptableAnswers: ['बुद्ध एयर', 'Buddha Air', 'बुद्ध एयरलाइन्स']
@@ -192,7 +184,7 @@ const logoData = {
       {
         id: 'tata',
         name: 'टाटा',
-        imagePath: '/logos/tata.png',
+        imagePath: '/yo-chineu/tata.png',
         difficulty: 'easy',
         category: 'सवारी साधन',
         acceptableAnswers: ['टाटा', 'Tata', 'टाटा मोटर्स', 'TATA']
@@ -200,7 +192,7 @@ const logoData = {
       {
         id: 'wai_wai',
         name: 'वाइ वाइ',
-        imagePath: '/logos/wai_wai.png',
+        imagePath: '/yo-chineu/wai_wai.png',
         difficulty: 'easy',
         category: 'खाद्य',
         acceptableAnswers: ['वाइ वाइ', 'Wai Wai', 'वाइवाइ', 'Wai-Wai']
@@ -208,7 +200,7 @@ const logoData = {
       {
         id: 'cg',
         name: 'सीजी',
-        imagePath: '/logos/cg.png',
+        imagePath: '/yo-chineu/cg.png',
         difficulty: 'medium',
         category: 'समूह',
         acceptableAnswers: ['सीजी', 'चौधरी समूह', 'CG', 'Chaudhary Group']
@@ -216,7 +208,7 @@ const logoData = {
       {
         id: 'dish_home',
         name: 'डिश होम',
-        imagePath: '/logos/dish_home.png',
+        imagePath: '/yo-chineu/dish_home.png',
         difficulty: 'medium',
         category: 'केबल',
         acceptableAnswers: ['डिश होम', 'Dish Home', 'डिशहोम']
@@ -224,7 +216,7 @@ const logoData = {
       {
         id: 'ntc',
         name: 'नेपाल टेलिकम',
-        imagePath: '/logos/ntc.png',
+        imagePath: '/yo-chineu/ntc.png',
         difficulty: 'easy',
         category: 'दूरसञ्चार',
         acceptableAnswers: ['नेपाल टेलिकम', 'NTC', 'नेपाल दूरसञ्चार']
@@ -232,7 +224,7 @@ const logoData = {
       {
         id: 'himalayan_bank',
         name: 'हिमालयन बैंक',
-        imagePath: '/logos/himalayan_bank.png',
+        imagePath: '/yo-chineu/himalayan_bank.png',
         difficulty: 'medium',
         category: 'बैंकिङ',
         acceptableAnswers: ['हिमालयन बैंक', 'HBL', 'हिमालयन बैंक लिमिटेड', 'Himalayan Bank']
@@ -240,7 +232,7 @@ const logoData = {
       {
         id: 'yeti_airlines',
         name: 'यति एयरलाइन्स',
-        imagePath: '/logos/yeti_airlines.png',
+        imagePath: '/yo-chineu/yeti_airlines.png',
         difficulty: 'medium',
         category: 'हवाई सेवा',
         acceptableAnswers: ['यति एयरलाइन्स', 'Yeti Airlines', 'यति']
@@ -248,7 +240,7 @@ const logoData = {
       {
         id: 'kumari_bank',
         name: 'कुमारी बैंक',
-        imagePath: '/logos/kumari_bank.png',
+        imagePath: '/yo-chineu/kumari_bank.png',
         difficulty: 'hard',
         category: 'बैंकिङ',
         acceptableAnswers: ['कुमारी बैंक', 'कुमारी', 'Kumari Bank', 'कुमारी बैंक लिमिटेड']
@@ -256,7 +248,7 @@ const logoData = {
       {
         id: 'laxmi_bank',
         name: 'लक्ष्मी बैंक',
-        imagePath: '/logos/laxmi_bank.png',
+        imagePath: '/yo-chineu/laxmi_bank.jpg',
         difficulty: 'hard',
         category: 'बैंकिङ',
         acceptableAnswers: ['लक्ष्मी बैंक', 'लक्ष्मी', 'Laxmi Bank']
@@ -264,7 +256,7 @@ const logoData = {
       {
         id: 'nic_asia',
         name: 'एनआईसी एशिया बैंक',
-        imagePath: '/logos/nic_asia.png',
+        imagePath: '/yo-chineu/nic_asia.png',
         difficulty: 'hard',
         category: 'बैंकिङ',
         acceptableAnswers: ['एनआईसी एशिया', 'एनआईसी एशिया बैंक', 'NIC Asia', 'NIC']
@@ -272,7 +264,7 @@ const logoData = {
       {
         id: 'nepal_investment_bank',
         name: 'नेपाल इन्भेस्टमेन्ट बैंक',
-        imagePath: '/logos/nepal_investment_bank.png',
+        imagePath: '/yo-chineu/nepal_investment_bank.png',
         difficulty: 'hard',
         category: 'बैंकिङ',
         acceptableAnswers: ['नेपाल इन्भेस्टमेन्ट बैंक', 'NIBL', 'नेपाल इन्भेस्टमेन्ट', 'Nepal Investment Bank']
@@ -280,7 +272,7 @@ const logoData = {
       {
         id: 'gorkha_brewery',
         name: 'गोरखा ब्रुअरी',
-        imagePath: '/logos/gorkha_brewery.png',
+        imagePath: '/yo-chineu/gorkha_brewery.png',
         difficulty: 'medium',
         category: 'पेय पदार्थ',
         acceptableAnswers: ['गोरखा ब्रुअरी', 'गोरखा', 'Gorkha Brewery', 'गोरखा बियर']
@@ -288,7 +280,7 @@ const logoData = {
       {
         id: 'nepal_airlines',
         name: 'नेपाल एयरलाइन्स',
-        imagePath: '/logos/nepal_airlines.png',
+        imagePath: '/yo-chineu/nepal_airlines.png',
         difficulty: 'easy',
         category: 'हवाई सेवा',
         acceptableAnswers: ['नेपाल एयरलाइन्स', 'NAC', 'नेपाल एयरलाइन्स कर्पोरेसन', 'Nepal Airlines', 'रोयल नेपाल एयरलाइन्स']
@@ -296,7 +288,7 @@ const logoData = {
       {
         id: 'esewa',
         name: 'ई-सेवा',
-        imagePath: '/logos/esewa.png',
+        imagePath: '/yo-chineu/esewa.jpg',
         difficulty: 'easy',
         category: 'फिन्टेक',
         acceptableAnswers: ['ई-सेवा', 'eSewa', 'ई सेवा', 'e-Sewa']
@@ -304,7 +296,7 @@ const logoData = {
       {
         id: 'khalti',
         name: 'खल्ती',
-        imagePath: '/logos/khalti.png',
+        imagePath: '/yo-chineu/khalti.png',
         difficulty: 'medium',
         category: 'फिन्टेक',
         acceptableAnswers: ['खल्ती', 'Khalti', 'खल्ती डिजिटल वालेट']
@@ -312,7 +304,7 @@ const logoData = {
       {
         id: 'worldlink',
         name: 'वर्ल्डलिङ्क',
-        imagePath: '/logos/worldlink.png',
+        imagePath: '/yo-chineu/worldlink.png',
         difficulty: 'easy',
         category: 'इन्टरनेट',
         acceptableAnswers: ['वर्ल्डलिङ्क', 'WorldLink', 'वर्ल्ड लिङ्क', 'वर्ल्डलिङ्क कम्युनिकेसन्स']
@@ -320,7 +312,7 @@ const logoData = {
       {
         id: 'siddhartha_bank',
         name: 'सिद्धार्थ बैंक',
-        imagePath: '/logos/siddhartha_bank.png',
+        imagePath: '/yo-chineu/siddhartha_bank.jpg',
         difficulty: 'hard',
         category: 'बैंकिङ',
         acceptableAnswers: ['सिद्धार्थ बैंक', 'सिद्धार्थ', 'Siddhartha Bank', 'सिद्धार्थ बैंक लिमिटेड']
